Extract a separate class for descriptive access
